requirements.txt
fastapi~=0.115.6
pytesseract~=0.3.13
easyocr~=1.7.2
uvicorn~=0.34.0
opencv-python~=4.11.0.86
python-multipart
numpy~=1.26.4
pillow~=11.1.0
requests
python-dotenv~=0.19.2